docs: categorize functions and make each function its own page
