+++ daft/functions/__init__.py
@@ -1,18 +1,19 @@
 from __future__ import annotations
 
-from .functions import (
-    monotonically_increasing_id,
+from .window import (
+    row_number,
+    rank,
+    dense_rank,
+)
+from .misc import monotonically_increasing_id, format
+from .columnar import (
     columns_sum,
     columns_mean,
     columns_avg,
     columns_min,
     columns_max,
-    row_number,
-    rank,
-    dense_rank,
-    format,
 )
-from .llm_generate import llm_generate
+from .llm import llm_generate
 
 __all__ = [
     "columns_avg",
