+++ docs/SUMMARY.md
@@ -1,0 +1,60 @@
+* Daft User Guide
+    * [Overview](index.md)
+    * [Installation](install.md)
+    * [Quickstart](quickstart.md)
+    * [Core Concepts](core_concepts.md)
+    * I/O
+        * [Overview](io/index.md)
+        * [Amazon Web Services](io/aws.md)
+        * [Apache Hudi](io/hudi.md)
+        * [Apache Iceberg](io/iceberg.md)
+        * [Delta Lake](io/delta_lake.md)
+        * [Hugging Face Datasets](io/huggingface.md)
+        * [Microsoft Azure](io/azure.md)
+        * [SQL](io/sql.md)
+    * [SQL](sql_overview.md)
+    * [Sessions](sessions.md)
+    * Catalogs
+        * [Overview](catalogs/index.md)
+        * [AWS Glue](catalogs/glue.md)
+        * [AWS S3 Tables](catalogs/s3tables.md)
+        * [Unity Catalog](catalogs/unity_catalog.md)
+    * [Spark Connect](spark_connect.md)
+    * [Distributed Computing](distributed.md)
+    * [Integrations](integrations.md)
+    * Advanced
+        * [Managing Memory Usage](advanced/memory.md)
+        * [Partitioning](advanced/partitioning.md)
+        * [Observability](advanced/observability.md)
+    * Resources
+        * [Architecture](resources/architecture.md)
+        * [Engine Comparison](resources/engine_comparison.md)
+        * [Tutorials](resources/tutorials.md)
+        * [Benchmarks](resources/benchmarks/tpch.md)
+        * [Telemetry](resources/telemetry.md)
+    * Migration Guide
+        * [Coming from Dask](migration/dask_migration.md)
+    * [Roadmap](roadmap.md)
+    * [Release Notes <sup>↗</sup>](https://github.com/Eventual-Inc/Daft/releases)
+* Python API
+    * [Overview](api/index.md)
+    * [I/O](api/io.md)
+    * [DataFrame](api/dataframe.md)
+    * [Expressions](api/expressions.md)
+    * [Functions](api/functions2/)
+    * [Sessions](api/sessions.md)
+    * [Catalogs & Tables](api/catalogs_tables.md)
+    * [Schema](api/schema.md)
+    * [Data Types](api/datatypes.md)
+    * [Aggregations](api/aggregations.md)
+    * [Series](api/series.md)
+    * [Configuration](api/config.md)
+    * [Miscellaneous](api/misc.md)
+* SQL Reference
+    * [Overview](sql/index.md)
+    * Statements
+        * [SELECT](sql/statements/select.md)
+        * [SHOW](sql/statements/show.md)
+        * [USE](sql/statements/use.md)
+    * [Data Types](sql/datatypes.md)
+    * [Window Functions](sql/window_functions.md)
