+++ docs/gen_pages/templates/function_page.md.j2
@@ -1,0 +1,7 @@
+---
+hide:
+  - toc
+---
+
+# {{ fn.path }}
+::: {{ fn.path }}
