+++ requirements-doc.txt
@@ -10,3 +10,7 @@ mkdocs-minify-plugin
 mkdocs-redirects
 yamllint
 yamlfix
+mkdocs-gen-files
+griffe
+mkdocs-literate-nav
+Jinja2
