+++ tests/functions/test_llm_generate.py
@@ -5,7 +5,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from daft import Expression, Series, lit
-from daft.functions.llm_generate import _vLLMGenerator, llm_generate
+from daft.functions.llm import _vLLMGenerator, llm_generate
 
 vllm = pytest.importorskip("vllm")
